requirements.txt
python-dotenv
numpy
pandas
pymongo[srv]
certifi
pyyaml
notebook
python-box
seaborn
matplotlib
scikit-learn
scipy
mlflow
streamlit